Switched int to uint8_t or uint16_t

+++ src/headers/LCD.hpp
@@ -11,9 +11,9 @@ class LCD : public MI0283QT9
 
     void write(const char *text, int_least16_t x, int_least16_t y);
 
-    void write(const char * text, int_least16_t x, int_least16_t y, int s);
+    void write(const char * text, int_least16_t x, int_least16_t y, uint8_t s);
 
-    void calibrate(int screen_width, int screen_height);
+    void calibrate(uint16_t screen_width, uint8_t screen_height);
 
     void setPage(uint8_t page);
